+++ internal/biz/namespace.go
@@ -3,29 +3,62 @@ package biz
 import (
 	"context"
 	"errors"
+	"strings"
+	"time"
 
+	"github.com/aide-family/magicbox/safety"
+	"github.com/aide-family/magicbox/strutil"
 	"github.com/bwmarrin/snowflake"
 	klog "github.com/go-kratos/kratos/v2/log"
 	"gorm.io/gorm"
 
 	"github.com/aide-family/rabbit/internal/biz/bo"
 	"github.com/aide-family/rabbit/internal/biz/repository"
+	"github.com/aide-family/rabbit/internal/biz/vobj"
+	"github.com/aide-family/rabbit/internal/conf"
 	"github.com/aide-family/rabbit/pkg/merr"
 )
 
-func NewNamespace(namespaceRepo repository.Namespace, helper *klog.Helper) *Namespace {
+func NewNamespace(
+	bc *conf.Bootstrap,
+	namespaceRepo repository.Namespace,
+	helper *klog.Helper,
+) *Namespace {
+	namespaces := safety.NewSyncMap(make(map[snowflake.ID]*bo.NamespaceItemBo))
+	for _, namespace := range conf.GetFileConfig().GetNamespaces() {
+		uid := snowflake.ParseInt64(namespace.GetUid())
+		createdAt, _ := time.Parse(time.DateTime, namespace.GetCreatedAt())
+		updatedAt, _ := time.Parse(time.DateTime, namespace.GetUpdatedAt())
+		// Note: Config_Namespace.GetName() returns int64 in proto, convert to string
+		name := namespace.GetName()
+		namespaces.Set(uid, &bo.NamespaceItemBo{
+			UID:       uid,
+			Name:      name,
+			Metadata:  namespace.GetMetadata(),
+			Status:    vobj.GlobalStatus(namespace.GetStatus()),
+			CreatedAt: createdAt,
+			UpdatedAt: updatedAt,
+		})
+	}
 	return &Namespace{
+		useDatabase:   bc.GetUseDatabase() == "true",
 		namespaceRepo: namespaceRepo,
+		namespaces:    namespaces,
 		helper:        klog.NewHelper(klog.With(helper.Logger(), "biz", "namespace")),
 	}
 }
 
 type Namespace struct {
 	helper        *klog.Helper
+	useDatabase   bool
 	namespaceRepo repository.Namespace
+	namespaces    *safety.SyncMap[snowflake.ID, *bo.NamespaceItemBo]
 }
 
 func (n *Namespace) CreateNamespace(ctx context.Context, req *bo.CreateNamespaceBo) error {
+	if !n.useDatabase {
+		return merr.ErrorParamsNotSupportFileConfig()
+	}
 	if _, err := n.namespaceRepo.GetNamespaceByName(ctx, req.Name); err == nil {
 		return merr.ErrorParams("namespace %s already exists", req.Name)
 	} else if !errors.Is(err, gorm.ErrRecordNotFound) {
@@ -40,6 +73,9 @@ func (n *Namespace) CreateNamespace(ctx context.Context, req *bo.CreateNamespace
 }
 
 func (n *Namespace) UpdateNamespace(ctx context.Context, req *bo.UpdateNamespaceBo) error {
+	if !n.useDatabase {
+		return merr.ErrorParamsNotSupportFileConfig()
+	}
 	doNamespace := req.ToDoNamespace()
 	existNamespace, err := n.namespaceRepo.GetNamespaceByName(ctx, doNamespace.Name)
 	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
@@ -56,6 +92,9 @@ func (n *Namespace) UpdateNamespace(ctx context.Context, req *bo.UpdateNamespace
 }
 
 func (n *Namespace) UpdateNamespaceStatus(ctx context.Context, req *bo.UpdateNamespaceStatusBo) error {
+	if !n.useDatabase {
+		return merr.ErrorParamsNotSupportFileConfig()
+	}
 	if err := n.namespaceRepo.UpdateNamespaceStatus(ctx, req); err != nil {
 		n.helper.Errorw("msg", "update namespace status failed", "error", err, "uid", req.UID)
 		return merr.ErrorInternal("update namespace status %s failed", req.UID)
@@ -64,6 +103,9 @@ func (n *Namespace) UpdateNamespaceStatus(ctx context.Context, req *bo.UpdateNam
 }
 
 func (n *Namespace) DeleteNamespace(ctx context.Context, uid snowflake.ID) error {
+	if !n.useDatabase {
+		return merr.ErrorParamsNotSupportFileConfig()
+	}
 	if err := n.namespaceRepo.DeleteNamespace(ctx, uid); err != nil {
 		n.helper.Errorw("msg", "delete namespace failed", "error", err, "uid", uid)
 		return merr.ErrorInternal("delete namespace %s failed", uid)
@@ -71,7 +113,18 @@ func (n *Namespace) DeleteNamespace(ctx context.Context, uid snowflake.ID) error
 	return nil
 }
 
+func (n *Namespace) getNamespaceByFileConfigWithUID(ctx context.Context, uid snowflake.ID) (*bo.NamespaceItemBo, error) {
+	namespace, ok := n.namespaces.Get(uid)
+	if !ok {
+		return nil, merr.ErrorNotFound("namespace %s not found", uid)
+	}
+	return namespace, nil
+}
+
 func (n *Namespace) GetNamespace(ctx context.Context, uid snowflake.ID) (*bo.NamespaceItemBo, error) {
+	if !n.useDatabase {
+		return n.getNamespaceByFileConfigWithUID(ctx, uid)
+	}
 	doNamespace, err := n.namespaceRepo.GetNamespace(ctx, uid)
 	if err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -84,7 +137,19 @@ func (n *Namespace) GetNamespace(ctx context.Context, uid snowflake.ID) (*bo.Nam
 	return bo.NewNamespaceItemBo(doNamespace), nil
 }
 
+func (n *Namespace) getNamespaceByFileConfigWithName(ctx context.Context, name string) (*bo.NamespaceItemBo, error) {
+	for _, namespace := range n.namespaces.Values() {
+		if namespace.Name == name {
+			return namespace, nil
+		}
+	}
+	return nil, merr.ErrorNotFound("namespace %s not found", name)
+}
+
 func (n *Namespace) GetNamespaceByName(ctx context.Context, name string) (*bo.NamespaceItemBo, error) {
+	if !n.useDatabase {
+		return n.getNamespaceByFileConfigWithName(ctx, name)
+	}
 	doNamespace, err := n.namespaceRepo.GetNamespaceByName(ctx, name)
 	if err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -96,7 +161,32 @@ func (n *Namespace) GetNamespaceByName(ctx context.Context, name string) (*bo.Na
 	return bo.NewNamespaceItemBo(doNamespace), nil
 }
 
+func (n *Namespace) getNamespaceByFileConfig(ctx context.Context, req *bo.ListNamespaceBo) ([]*bo.NamespaceItemBo, error) {
+	namespaces := make([]*bo.NamespaceItemBo, 0, n.namespaces.Len())
+	for _, namespace := range n.namespaces.Values() {
+		if strutil.IsNotEmpty(req.Keyword) && !strings.Contains(namespace.Name, req.Keyword) {
+			continue
+		}
+		if req.Status.Exist() && !req.Status.IsUnknown() && namespace.Status != req.Status {
+			continue
+		}
+		namespaces = append(namespaces, namespace)
+	}
+	total := int64(len(namespaces))
+	pageRequestBo := bo.NewPageRequestBo(req.Page, req.PageSize)
+	pageRequestBo.WithTotal(total)
+	req.PageRequestBo = pageRequestBo
+	return namespaces, nil
+}
+
 func (n *Namespace) ListNamespace(ctx context.Context, req *bo.ListNamespaceBo) (*bo.PageResponseBo[*bo.NamespaceItemBo], error) {
+	if !n.useDatabase {
+		namespaces, err := n.getNamespaceByFileConfig(ctx, req)
+		if err != nil {
+			return nil, err
+		}
+		return bo.NewPageResponseBo(req.PageRequestBo, namespaces), nil
+	}
 	pageResponseBo, err := n.namespaceRepo.ListNamespace(ctx, req)
 	if err != nil {
 		n.helper.Errorw("msg", "list namespace failed", "error", err, "req", req)
@@ -109,7 +199,41 @@ func (n *Namespace) ListNamespace(ctx context.Context, req *bo.ListNamespaceBo) 
 	return bo.NewPageResponseBo(pageResponseBo.PageRequestBo, items), nil
 }
 
+func (n *Namespace) getSelectNamespaceByFileConfig(ctx context.Context, req *bo.SelectNamespaceBo) ([]*bo.NamespaceItemSelectBo, error) {
+	namespaces := make([]*bo.NamespaceItemSelectBo, 0, n.namespaces.Len())
+	for _, namespace := range n.namespaces.Values() {
+		if strutil.IsNotEmpty(req.Keyword) && !strings.Contains(namespace.Name, req.Keyword) {
+			continue
+		}
+		if req.Status.Exist() && !req.Status.IsUnknown() && namespace.Status != req.Status {
+			continue
+		}
+		namespaces = append(namespaces, &bo.NamespaceItemSelectBo{
+			UID:      namespace.UID,
+			Name:     namespace.Name,
+			Status:   namespace.Status,
+			Disabled: namespace.Status != vobj.GlobalStatusEnabled,
+			Tooltip:  "",
+		})
+	}
+	total := int64(len(namespaces))
+	req.Limit = int32(total)
+	req.LastUID = 0
+	return namespaces, nil
+}
+
 func (n *Namespace) SelectNamespace(ctx context.Context, req *bo.SelectNamespaceBo) (*bo.SelectNamespaceBoResult, error) {
+	if !n.useDatabase {
+		namespaces, err := n.getSelectNamespaceByFileConfig(ctx, req)
+		if err != nil {
+			return nil, err
+		}
+		return &bo.SelectNamespaceBoResult{
+			Items:   namespaces,
+			Total:   int64(len(namespaces)),
+			LastUID: 0,
+		}, nil
+	}
 	result, err := n.namespaceRepo.SelectNamespace(ctx, req)
 	if err != nil {
 		n.helper.Errorw("msg", "select namespace failed", "error", err, "req", req)
